feat: creating account pg+login box component
